refactor[generated]: Move `Attribute` to a separate module.
